requirements.txt
Flask==1.1.1
Flask-Restful==0.3.8
Flask-SQLAlchemy==2.4.1